moved walletAddToken to frontend/src/utils/
